refactor(step10): 예약, 오류 controller 분리
