+++ src/main/java/roomescape/controller/ExceptionController.java
@@ -1,0 +1,12 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+
+public class ExceptionController {
+
+    @ExceptionHandler(RuntimeException.class)
+    public ResponseEntity handleException(RuntimeException e) {
+        return ResponseEntity.badRequest().build();
+    }
+}
